Resolves 29 conflicted files from merging the lint-first selector adoption
(N/SIM/RUF/ERA/T20 + pyright strict) onto top of 105 local-ahead commits
covering M07-M12 (render orchestration, publish-to-HF, audit pipeline,
recipe lint, validation drift guards, paragraph alignment, glyph emission).

Resolution strategy per file type:
- add/add conflicts (20 files): kept local M07-M12 version as canonical;
  applied origin's lint fixes on top post-merge.
- content conflicts (9 files): took origin's side (lint-first simplifications:
  SIM108 ternaries, contextlib.suppress, next(iter(...)), !s format, raw
  regex strings, unicode escapes).

Post-resolution lint fixes applied to local-side files:
- RUF046: removed redundant int() wrapping around round() in render/line,
  page, paragraph, publish/dataset_card
- SIM108: converted if/else blocks to ternary in render/page, paragraph,
  publish/dataset_card
- RUF005: [*current, word] spread in render/wrap
- RUF007: itertools.pairwise() in test_render_{line,page,paragraph},
  test_cli_render_pages
- N802: renamed uppercase test functions in test_lint; added noqa:N802
  for name-embedding functions in test_spec_docs
- N818: renamed _Boom -> _BoomError in test_audit
- N811: renamed import alias to _PUBLISH_TRUTHY in test_trainer_dataset_contract
- N814: renamed _RC -> _RenderCtx import alias in test_render_page
- ERA001: rewrote code-looking comment in test_render_paragraph
- RUF043: escaped regex patterns (labels\.json, output\.mode) in test_publish_*
- RUF002/RUF003: replaced × and ∪ with ascii equivalents or noqa
- RUF100: removed stale noqa directives

+++ src/pd_ocr_synth/render/word_crop.py
@@ -93,7 +93,7 @@ def render_word_crop(
         raise MissingGlyphError(text, font.path, missing)
 
     handles = ctx.font_handles(font.path)
-    pixel_size = max(1, int(round(font_size_pt * dpi / 72.0)))
+    pixel_size = max(1, round(font_size_pt * dpi / 72.0))
     handles.ft_face.set_pixel_sizes(pixel_size, pixel_size)
 
     info_glyphs, positions = _shape(handles.hb_face, text, pixel_size, font.features)
@@ -234,7 +234,7 @@ def _composite(
         # baseline-to-top distance, so glyph top edge is pen_y -
         # bitmap_top (with HarfBuzz y-offsets in font coords).
         y = pen_y - y_offset - bm["top"]
-        placements.append((bm, int(round(x)), int(round(y)), info.cluster))
+        placements.append((bm, round(x), round(y), info.cluster))
         pen_x += pos.x_advance / 64.0
         pen_y += pos.y_advance / 64.0
